requirements.txt
django-allauth==0.42.0
django-anymail==7.2.1
django-bootstrap4==2.2.0
django-countries==6.1.3
django-html-emailer @ git+https://github.com/if-then-fund/django-html-emailer@9a31d8ee5ba44d1820a42700ff2002f9f705808c
django-localflavor==3.0.1
django-picklefield==3.0.1
django-registration==3.1
django-thumbs-v2==0.4.1
django-widget-tweaks==1.4.8